Refactors PokeApi handling

+++ Backend/src/main/java/com/codecool/backend/repository/pokemon/PokemonAssetRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 
 @Repository
 public interface PokemonAssetRepository extends JpaRepository<PokemonAsset, Integer> {
-    public Optional<PokemonAsset> findBySpeciesId(Long speciesId);
+    public Optional<PokemonAsset> findBySpeciesId(Integer speciesId);
 
 }
